separates the message class from the stream class file
updates patient_zero.rb to include message
